moved dir sub-folder moved up one level
