+++ src/sub-components/additional-section/styled-additional-section.js
