+++ src/sub-components/footer-content/styled-footer-content.js
@@ -1,5 +1,5 @@
 import { styled } from "linaria/react";
-import { device } from "../../../../components/utils/devices";
+import { device } from "../../../components/utils/devices";
 
 const StyledFooter = styled.div`
   background: ${(props) => (props.footerHomePage ? "#f9f9f9" : "transparent")};
