+++ src/sub-components/head/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Helmet } from "react-helmet";
-import languages from "../../../../languages";
+import languages from "../../../languages";
 
 const Head = (props) => {
   const { metaDescription, metaKeywords, title, metaDescriptionOg } = props;
