+++ src/sub-components/header-content/index.js
@@ -2,10 +2,10 @@ import React from "react";
 import { Link } from "gatsby";
 
 import StyledHeaderContent from "./styled-header-content";
-import LanguageSelector from "../../../../components/language-selector";
-import HeaderLogo from "../../../../static/icons/header-logo.react.svg";
-import Button from "../../../../components/button";
-import Text from "../../../../components/text";
+import LanguageSelector from "../../../components/language-selector";
+import HeaderLogo from "../../../static/icons/header-logo.react.svg";
+import Button from "../../../components/button";
+import Text from "../../../components/text";
 
 const HeaderContent = (props) => {
   const { t, language, href, labelButton, headerText, ...rest } = props;
