+++ src/sub-components/header-content/styled-header-content.js
