+++ src/sub-components/license/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 
-import Text from "../../../../components/text";
-import Link from "../../../../components/link";
+import Text from "../../../components/text";
+import Link from "../../../components/link";
 
 const License = (props) => {
   const { t } = props;
